reorganize provider module (#267)
